Move many chat functions to separate helper
